Introduced shell.confirm()
Use it in clean job.

+++ Sources/Devops/Jobs/CleanJob.swift
@@ -1,13 +1,13 @@
 import Foundation
 import Common
 
-// TODO: Ask confirmation!
-
 public class CleanJob: DevopsJob {
     public override func run() {
         do {
-            try shell.removeFile(atPath: "Dockerfile")
-            try shell.removeDirectory(atPath: "chart")
+            if shell.confirm("Are you sure you want to delete all generated files?") {
+                try shell.removeFile(atPath: "Dockerfile")
+                try shell.removeDirectory(atPath: "chart")
+            }
         } catch {
             print(error.localizedDescription)
         }
